turn creatId into long type

+++ localfullnode2/src/main/java/one/inve/localfullnode2/dep/items/CreatorId.java
@@ -3,13 +3,13 @@ package one.inve.localfullnode2.dep.items;
 import one.inve.localfullnode2.dep.DependentItem;
 
 public class CreatorId extends DependentItem {
-	private int creatorId;
+	private long creatorId;
 
-	public int get() {
+	public long get() {
 		return creatorId;
 	}
 
-	public void set(int creatorId) {
+	public void set(long creatorId) {
 		this.creatorId = creatorId;
 	}
